+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 2.19 branch
 ### 2.19.11
 * fixed little bug for `HtmlHelper::image()` method with `$path` as array;
+* fixed little bug for `HtmlHelper::link()` method with `$title` as array;
 * updated for `php-tools` 1.4.7.
 
 ### 2.19.10